change output_parser vars and renamed codeoutputparser
